Update hooks for azd 0.8.0
azd version 0.8.0 packs the application before creating azure resources (when running azd up)
Then `azd deploy` use the output from that packing to publish the app to Azure.
The `predeploy` hook is not valid anymore on this scenario to include the frontend to the backend. Instead, we need to switch to use `prepackage` for the backend service, so the frontend is attached before creating the package.